+++ astropy/units/cds.py
@@ -172,13 +172,15 @@ def enable():
     """
     Enable CDS units so they appear in results of
     `~astropy.units.UnitBase.find_equivalent_units` and
-    `~astropy.units.UnitBase.compose`.
+    `~astropy.units.UnitBase.compose`.  This will disable
+    all of the "default" `astropy.units` units, since there
+    are some namespace clashes between the two.
 
     This may be used with the ``with`` statement to enable CDS
     units only temporarily.
     """
     # Local import to avoid cyclical import
-    from .core import add_enabled_units
+    from .core import set_enabled_units
     # Local import to avoid polluting namespace
     import inspect
-    return add_enabled_units(inspect.getmodule(enable))
+    return set_enabled_units(inspect.getmodule(enable))