+++ astropy/units/tests/test_units.py
@@ -688,3 +688,15 @@ def test_fractional_rounding_errors_simple():
     assert isinstance(x.powers[0], Fraction)
     assert x.powers[0].numerator == 6
     assert x.powers[0].denominator == 5
+
+
+def test_enable_unit_groupings():
+    from ...units import cds
+    assert cds.geoMass not in u.kg.find_equivalent_units()
+
+    with cds.enable():
+        assert cds.geoMass in u.kg.find_equivalent_units()
+
+    from ...units import imperial
+    with imperial.enable():
+        assert imperial.inch in u.m.find_equivalent_units()